backend/lambdas/layers/pyjwt/requirements.txt
PyJWT==2.10.1
python-jose
requests